+++ business-easy-core/src/main/java/com/businesseasy/core/repositories/UnitConversionRepository.java
@@ -3,5 +3,5 @@ package com.businesseasy.core.repositories;
 import com.businesseasy.core.entities.UnitConversionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnitConversionRepository extends JpaRepository<UnitConversionEntity, Integer> {
+public interface UnitConversionRepository extends JpaRepository<UnitConversionEntity, Long> {
 }
