+++ business-easy-core/src/main/java/com/businesseasy/core/repositories/UnitRepository.java
@@ -3,5 +3,5 @@ package com.businesseasy.core.repositories;
 import com.businesseasy.core.entities.UnitEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnitRepository extends JpaRepository<UnitEntity, Integer> {
+public interface UnitRepository extends JpaRepository<UnitEntity, Long> {
 }
